Add basic module/plugin system
"Plugins" are currently referred to as modules, which is a little
confusing in the context of Python, so the name may be changed in the
future.

Currently Cinnabar modules are specified as a module path, and any
subclasses of Module found in the Python module will be instantiated in
the specified position.